k8s_prompt/pyproject.toml
[tool.poetry]
name = "k8s-prompt"
version = "0.1.0"
description = ""
authors = ["Carlo Scarioni <carlo.scarioni@mettle.co.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
#langchain = "0.1.20"
langchain-community = "0.2.0rc1"
langchain-openai = "*"
openai = "1.28.2"
python-dotenv = "1.0.1"
prompt_toolkit = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
